% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbc4va_data.R
\docType{data}
\name{nbc4vaDataRaw}
\alias{nbc4vaDataRaw}
\title{Example of unclean data in nbc4va}
\format{A dataframe with 100 rows and 102 columns:
\itemize{
  \item id (vectorof char): the case identifiers
  \item cause (vectorof char): the cause of death for each case
  \item symptom1..100 (vectorsof (1 OR 0 OR 99)): whether the symptom is recorded as present (1), absent (0), or unknown (99) for each case (row)
  \item Example:
  \tabular{ccccc}{
    id \tab cause \tab symptom1 \tab symptom2 \tab symptom3 \cr
    "a27" \tab "cause10" \tab 99 \tab 0 \tab 1\cr
    "k37" \tab "cause2" \tab 0 \tab 99 \tab 1\cr
    "e57" \tab "cause8" \tab 1 \tab 0 \tab 99
  }
}}
\source{
Random generation using the \code{\link{sample}} function with \code{\link{set.seed}} set to 1.
}
\usage{
nbc4vaDataRaw
}
\description{
A random generation of unclean verbal autopsy synthetic data for use in demonstrating the nbc4va package.
}
\details{
\strong{Warning}: This data may produce errors depending on how you use it in the package.
}
\examples{
library(nbc4va)
data(nbc4vaDataRaw)
}
\keyword{datasets}
