% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbc4va_help.R
\name{nbc4vaHelp}
\alias{nbc4vaHelp}
\title{Help functions in nbc4va}
\usage{
nbc4vaHelp()
}
\description{
Help functions in nbc4va
}
\section{About}{

This documentation page provides a list of the available help pages in the nbc4va package.
}

\section{Available Help Documentation}{

The nbc4va package has several help functions built in to aid the understanding of the methodology
of the package and to guide the use of functions and methods. It is possible to click the links to
the functions below to view the help pages or call them which produces the same result.
\itemize{
  \item \code{\link{nbc4vaHelp}}: opens this help page
  \item \code{\link{nbc4vaHelpMethods}}: opens the help page for methodology and implementation
  \item \code{\link{nbc4vaHelpData}}: opens the help page for formatting input training/testing data
  \item \code{\link{nbc4vaHelpBasic}}: opens the help page for using the package in graphical or file interface
  \item \code{\link{nbc4vaHelpAdvanced}}: opens the help page for using the package in R
  \item \code{\link{nbc4vaHelpFunctions}}: opens the help page for an organized list of useful functions in the package
  \item \code{\link{nbc4vaHelpDev}}: opens the developer help page for function and package dependencies
}
\subsection{Sample Code for Opening Help Documentation}{
  Run the following code using \code{\link{nbc4vaHelp}} in the R console to conveniently open up a variety of help pages.
  \preformatted{
--------------------------------------------------------------------------------------------------

    library(nbc4va)  # load the nbc4va package
    nbc4vaHelp()  # view this help page
    nbc4vaHelpMethods()  # view methods and implementation help page
    nbc4vaHelpData()  # view input train/test data formatting  help page
    nbc4vaHelpBasic()  # view graphical or file interface help page
    nbc4vaHelpAdvanced()  # view usage in R help page
    nbc4vaHelpFunctions()  # view help page of useful functions
    nbc4vaHelpDev()  # view developer help page

--------------------------------------------------------------------------------------------------
  }
}
}
\seealso{
Guide for package: \code{\link{nbc4va}}

Other help functions: \code{\link{nbc4vaHelpAdvanced}},
  \code{\link{nbc4vaHelpBasic}},
  \code{\link{nbc4vaHelpData}},
  \code{\link{nbc4vaHelpDev}},
  \code{\link{nbc4vaHelpFunctions}},
  \code{\link{nbc4vaHelpMethods}}
}

